feat(onboarding): drop the Done step from the sharing path
Continue from the last sharing step (CLI install) now seeds the welcome
page and redirects directly to /workspaces/{id}. No 'You're set up' card,
no Done chip in the progress bar.

- Extracted seedWelcomePage out of DoneStep into a standalone module so
  the wizard can call it without rendering Done UI.
- PathDef.doneStep is now optional. When omitted, the wizard's nextStep
  from the last step runs the seed + redirect.
- ProgressBar accepts showDone and skips the Done chip when false.
- Migrant and Memory paths keep their Done UI; only Sharing skips it.
- While in there, readSource() now recognizes 'drive' (was missing
  since the Drive source got added).
